fix: refactor usages
